Get ByteBuffer from the buffer wrapped inside
UnscopedBuffer wraps another buffer. So its ByteBuffer view
returned from getByteBuffer() should be either read-only or writable
depending on the nature of the buffer wrapped inside.

But it was broken since its getByteBuffer() invokes
DirectUtils.wrapAddress() directly inside and ignoring the writability
of the wrapped buffer.

The problem is fixed by delegating to the wrapped buffer's
getByteBuffer() instead.